refactoring server
