+++ server/helpers/index.js
@@ -1,0 +1,5 @@
+const momentToLong = require('./moment-to-long');
+
+module.exports = {
+  momentToLong,
+};
