+++ server/middleware/tchannel-client/helpers/get-peers.js
@@ -1,0 +1,5 @@
+const getPeers = () => process.env.CADENCE_TCHANNEL_PEERS
+  ? process.env.CADENCE_TCHANNEL_PEERS.split(',')
+  : ['127.0.0.1:7933'];
+
+module.exports = getPeers;
