+++ server/middleware/tchannel-client/helpers/index.js
@@ -1,0 +1,13 @@
+const cliTransform = require('./cli-transform');
+const getPeers = require('./get-peers');
+const lookupAsync = require('./lookup-async');
+const makeChannel = require('./make-channel');
+const uiTransform = require('./ui-transform');
+
+module.exports = {
+  cliTransform,
+  getPeers,
+  lookupAsync,
+  makeChannel,
+  uiTransform,
+};
