+++ server/middleware/tchannel-client/index.js
@@ -21,130 +21,14 @@
 
 'use strict';
 
-const path = require('path'),
-  dns = require('dns'),
-  get = require('lodash.get'),
-  TChannelAsThrift = require('tchannel/as/thrift'),
-  TChannel = require('tchannel'),
-  Long = require('long'),
-  losslessJSON = require('lossless-json'),
-  moment = require('moment'),
-  isIPv4 = require('is-ipv4-node');
+const get = require('lodash.get');
+const TChannel = require('tchannel');
 
-function uiTransform(item) {
-  if (!item || typeof item !== 'object') {
-    return item;
-  }
-
-  Object.entries(item).forEach(([subkey, subvalue]) => {
-    if (subvalue && typeof subvalue.unsigned === 'boolean') {
-      item[subkey] = Long.fromValue(subvalue).toNumber();
-      const m = moment(item[subkey] / 1000000);
-
-      if (m.isValid() && m.isAfter('2017-01-01')) {
-        item[subkey] = m.toISOString();
-      }
-    } else if (Buffer.isBuffer(subvalue)) {
-      if (subkey === 'nextPageToken') {
-        item.nextPageToken = subvalue.toString('base64');
-
-        return;
-      }
-
-      const stringval = subvalue.toString('utf8');
-
-      try {
-        // most of Cadence's uses of buffer is just line-delimited JSON.
-        item[subkey] = stringval
-          .split('\n')
-          .filter(x => x)
-          .map(JSON.parse);
-
-        if (item[subkey].length === 1) {
-          item[subkey] = item[subkey][0];
-        }
-      } catch (e) {
-        item[`${subkey}_base64`] = subvalue.toString('base64');
-        item[subkey] = stringval;
-      }
-    } else if (Array.isArray(subvalue)) {
-      subvalue.forEach(uiTransform);
-    } else if (subvalue && typeof subvalue === 'object') {
-      uiTransform(subvalue);
-    }
-  });
-
-  return item;
-}
-
-function cliTransform(item) {
-  if (!item || typeof item !== 'object') {
-    return item;
-  }
-
-  Object.entries(item).forEach(([subkey, subvalue]) => {
-    if (subvalue && typeof subvalue.unsigned === 'boolean') {
-      item[subkey] = new losslessJSON.LosslessNumber(
-        Long.fromValue(subvalue).toString()
-      );
-    } else if (Buffer.isBuffer(subvalue)) {
-      item[subkey] = subvalue.toString('base64');
-    } else if (Array.isArray(subvalue)) {
-      subvalue.forEach(cliTransform);
-    } else if (subvalue && typeof subvalue === 'object') {
-      cliTransform(subvalue);
-    } else if (subvalue === null || subvalue === undefined) {
-      delete item[subkey];
-    }
-  });
-
-  return item;
-}
-
-const lookupAsync = host =>
-  new Promise(function(resolve, reject) {
-    dns.lookup(host, { family: 4 }, function(err, ip) {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(ip);
-      }
-    });
-  });
-
-const peers = process.env.CADENCE_TCHANNEL_PEERS
-  ? process.env.CADENCE_TCHANNEL_PEERS.split(',')
-  : ['127.0.0.1:7933'];
-
-async function makeChannel(client) {
-  const ipPeers = await Promise.all(
-    peers.map(peer => {
-      const [host, port] = peer.split(':');
-
-      if (!isIPv4(host)) {
-        return lookupAsync(host).then(ip => [ip, port].join(':'));
-      } else {
-        return peer;
-      }
-    })
-  );
-
-  const cadenceChannel = client.makeSubChannel({
-    serviceName: 'cadence-frontend',
-    peers: ipPeers,
-    requestDefaults: {
-      hasNoParent: true,
-      headers: { as: 'raw', cn: 'cadence-web' },
-    },
-  });
-
-  const tchannelAsThrift = TChannelAsThrift({
-    channel: cadenceChannel,
-    entryPoint: path.join(__dirname, '../idl/cadence.thrift'),
-  });
-
-  return tchannelAsThrift;
-}
+const {
+  cliTransform,
+  makeChannel,
+  uiTransform,
+} = require('./helpers');
 
 module.exports = async function(ctx, next) {
   const client = TChannel();
